heatercontrol/requirements.txt
RPi.Gpio
paho-mqtt